admin

+++ frontend/src/components/AdminDashboard/AdminDashboard.jsx
@@ -1,5 +1,8 @@
 import React, { useEffect, useState } from 'react';
+import Calendar from 'react-calendar';
 import './AdminDashboard.css';
+import {formatSelectedDate, generateTimeSlots, durations, isOverlapping } from '../../utils/calendar';
+// import moment from './moment'
 
 const AdminDashboard = ({ token }) => {
   const [reservations, setReservations] = useState([]);
@@ -7,13 +10,47 @@ const AdminDashboard = ({ token }) => {
   const [coupons, setCoupons] = useState([]); // State for coupons
   const [error, setError] = useState('');
   const [isLoading, setIsLoading] = useState(true);
-  const [selectedDate, setSelectedDate] = useState('');
+  const [selectedDate, setSelectedDate] = useState(''); // Calendar selected date
   const [activeTab, setActiveTab] = useState('bookings'); // State to track active tab
+  const [expandedRows, setExpandedRows] = useState([]); // Track which rows are expanded
+
+    // Helper function to get today's date in YYYY-MM-DD format
+    const getCurrentDate = () => {
+      const today = new Date();
+      const year = today.getFullYear();
+      const month = String(today.getMonth() + 1).padStart(2, '0'); // Ensure two digits for month
+      const day = String(today.getDate()).padStart(2, '0'); // Ensure two digits for day
+      return `${year}-${month}-${day}`;
+    };
 
   // New states for creating a reservation
-  const [newDate, setNewDate] = useState('');
+  const [newDate, setNewDate] = useState(getCurrentDate())
   const [newTime, setNewTime] = useState('');
   const [newDuration, setNewDuration] = useState('');
+  const [newFirstName, setNewFirstName] = useState('');
+  const [newLastName, setNewLastName] = useState('');
+  const [newEmail, setNewEmail] = useState('');
+  const [newPhone, setNewPhone] = useState('');
+  const [finalPrice, setFinalPrice] = useState(0);
+
+  const formatDate = (date) => {
+    if (!date) return '';
+    
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0'); // Months are 0-based
+    const day = String(date.getDate()).padStart(2, '0');
+  
+    return `${year}-${month}-${day}`; // Local format: YYYY-MM-DD
+  };
+
+    // Toggle row expansion to show/hide details
+    const toggleRow = (id) => {
+      setExpandedRows((prevExpandedRows) =>
+        prevExpandedRows.includes(id)
+          ? prevExpandedRows.filter((rowId) => rowId !== id) // Remove if already expanded
+          : [...prevExpandedRows, id] // Add row to expanded
+      );
+    };
 
   // Fetch Reservations when bookings tab is active
   useEffect(() => {
@@ -63,22 +100,27 @@ const AdminDashboard = ({ token }) => {
     }
   }, [token, activeTab]);
 
+  const formattedDate = formatDate(selectedDate); // Format the selected date
   // Filter Reservations by Date
   useEffect(() => {
     if (selectedDate && activeTab === 'bookings') {
       const filtered = reservations
-        .filter((reservation) => reservation.date === selectedDate)
+        .filter((reservation) => {
+          const reservationDate = new Date(reservation.date).toISOString().split('T')[0]; // Format reservation date
+          return reservationDate === formattedDate;
+        })
         .sort((a, b) => new Date(a.date + 'T' + a.time) - new Date(b.date + 'T' + b.time));
       setFilteredReservations(filtered);
     } else {
-      const sortedReservations = [...reservations].sort(
-        (a, b) => new Date(a.date + 'T' + a.time) - new Date(b.date + 'T' + b.time)
-      );
-      setFilteredReservations(sortedReservations);
+      setFilteredReservations(reservations);
     }
   }, [selectedDate, reservations, activeTab]);
 
   const handleCreateReservation = async () => {
+    if (isOverlapping(newTime, newDuration, new Date(newDate), reservations)) {
+      window.alert('Selected time slot overlaps with an existing reservation. Please choose a different time.');
+      return;
+    }
     try {
       const response = await fetch('http://localhost:3000/api/reservations', {
         method: 'POST',
@@ -86,31 +128,57 @@ const AdminDashboard = ({ token }) => {
           'Content-Type': 'application/json',
           Authorization: `Bearer ${token}`,
         },
-        body: JSON.stringify({ date: newDate, time: newTime, duration: parseInt(newDuration) }),
+        body: JSON.stringify({
+          date: newDate,
+          time: newTime,
+          duration: parseInt(newDuration),
+          firstName: newFirstName,
+          lastName: newLastName,
+          email: newEmail,
+          phone: newPhone,
+          finalPrice: finalPrice > 0 ? finalPrice : 1, // Ensure finalPrice is greater than 0
+        }),
       });
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status} ${response.statusText}`);
+  
+      const data = await response.json();
+  
+      if (response.ok) {
+        // Display success alert
+        window.alert('Reservation created successfully!');
+  
+        // Clear the form inputs by resetting state
+        setNewDate('');
+        setNewTime('');
+        setNewDuration('');
+        setNewFirstName('');
+        setNewLastName('');
+        setNewEmail('');
+        setNewPhone('');
+        setFinalPrice(0);
+        // setReservations([...reservations, data.newReservation]);
+        // Optionally, set a success message in state if you want to display it in the UI
+        // setSuccessMessage('Reservation created successfully!');
+      } else {
+        throw new Error(`Error: ${data.message}`);
       }
-
-      const newReservation = await response.json();
-      setReservations([...reservations, newReservation.newReservation]);
-      setError('');
-      setNewDate('');
-      setNewTime('');
-      setNewDuration('');
     } catch (err) {
-      setError(err.message);
+      console.error('Error creating reservation:', err.message);
     }
   };
 
   const handleDeleteReservation = async (id) => {
+    const confirmDelete = window.confirm('Are you sure you want to delete this reservation?');
+    
+    if (!confirmDelete) {
+      return; // If the user cancels, do nothing
+    }
+  
     try {
       const response = await fetch(`http://localhost:3000/api/reservations/${id}`, {
         method: 'DELETE',
         headers: { Authorization: `Bearer ${token}` },
       });
-
+  
       if (response.ok) {
         setReservations((prev) => prev.filter((reservation) => reservation._id !== id));
       } else {
@@ -121,8 +189,14 @@ const AdminDashboard = ({ token }) => {
       setError('Failed to delete the reservation. Please try again.');
     }
   };
-
+  
   const handleDeleteCoupon = async (couponCode) => {
+    const confirmDelete = window.confirm('Are you sure you want to delete this coupon?');
+    
+    if (!confirmDelete) {
+      return; // If the user cancels, do nothing
+    }
+  
     try {
       const response = await fetch('http://localhost:3000/api/coupons/delete', {
         method: 'POST',
@@ -132,7 +206,7 @@ const AdminDashboard = ({ token }) => {
         },
         body: JSON.stringify({ couponCode }),
       });
-
+  
       if (response.ok) {
         setCoupons((prev) => prev.filter((coupon) => coupon.code !== couponCode));
       } else {
@@ -144,6 +218,7 @@ const AdminDashboard = ({ token }) => {
     }
   };
 
+  
   return (
     <section className="admin-dashboard">
       {/* Tabs for switching between Bookings and Coupons */}
@@ -168,105 +243,176 @@ const AdminDashboard = ({ token }) => {
       {/* Conditional rendering based on active tab */}
       {activeTab === 'bookings' ? (
         <>
-          {/* Date Filter Input */}
-          <div className="admin-dashboard__container">
-            <label htmlFor="admin-dashboard__filter">Filter by Date: </label>
-            <input
-              type="date"
-              id="filter-date"
-              value={selectedDate}
-              onChange={(e) => setSelectedDate(e.target.value)}
-            />
-          </div>
+         {/* Calendar for selecting a date */}
+         <div className="admin-dashboard__container">
+          <div className="time__calendar-container">
+            <div className="time__calendar">
+              <Calendar
+                onClickDay={(date) => setSelectedDate(date)} 
+                value={selectedDate}
+              />
+            </div>
+            </div>
+            {/* </div> */}
 
-          {/* New Reservation Form */}
+            {isLoading ? (
+              <p>Loading reservations...</p>
+            ) : (
+              <>
+              <div>
+              <table className="admin-dashboard__table">
+                <thead>
+                  <tr>
+                    <th>Date</th>
+                    <th>Time</th>
+                    <th>Duration</th>
+                    <th>Price</th>
+                    <th>Details</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {filteredReservations.length > 0 ? (
+                    filteredReservations.sort((a, b) => {
+                      const dateA = new Date(`${a.date}T${a.time}`);
+                      const dateB = new Date(`${b.date}T${b.time}`);
+                      return dateA - dateB;
+                    }).map((reservation) => (
+                      <React.Fragment key={reservation._id}>
+                      <tr>
+                        <td>{new Date(reservation.date).toLocaleDateString()}</td>
+                        <td>{reservation.time}</td>
+                        <td>{reservation.duration} mins</td>
+                        <td>£{reservation.finalPrice}</td>
+                        <td>
+                          <button
+                            className="admin-dashboard__arrow"
+                            onClick={() => toggleRow(reservation._id)}
+                          >
+                            {expandedRows.includes(reservation._id) ? '▲' : '▼'}
+                          </button>
+                          <button
+                            className="admin-dashboard__button"
+                            onClick={() => handleDeleteReservation(reservation._id)}
+                          >
+                            Delete
+                          </button>
+                        </td>
+                      </tr>
+                      {expandedRows.includes(reservation._id) && (
+                        <tr>
+                          <td colSpan="5">
+                            <div className="reservation-details">
+                              <p><strong>First Name:</strong> {reservation.firstName}</p>
+                              <p><strong>Last Name:</strong> {reservation.lastName}</p>
+                              <p><strong>Email:</strong> {reservation.email}</p>
+                              <p><strong>Phone:</strong> {reservation.phone}</p>
+                            </div>
+                          </td>
+                        </tr>
+                      )}
+                    </React.Fragment>
+                    ))
+                  ) : (
+                    <tr>
+                      <td colSpan="5">No reservations found for the selected date.</td>
+                    </tr>
+                  )}
+                </tbody>
+              </table>
+                  {/* Add Reservation Form */}
+            </div>
+              </>
+            )}
+          </div>
           <div className="admin-dashboard__new-reservation">
-            <h3>Create New Reservation</h3>
-            <label>
-              Date:
-              <input
-                type="date"
-                value={newDate}
-                onChange={(e) => setNewDate(e.target.value)}
-                required
-              />
-            </label>
-            <label>
-              Time:
-              <input
-                type="time"
-                value={newTime}
-                onChange={(e) => setNewTime(e.target.value)}
-                required
-              />
-            </label>
-            <label>
+              <h3 className="admin-dashboard__title">Add New Reservation</h3>
+              <label className="admin-dashboard__new-reservation-label">
+                Date:
+                <input
+                className="admin-dashboard__new-reservation-input"
+                  type="date"
+                  value={newDate}
+                  onChange={(e) => setNewDate(e.target.value)}
+                  required
+                />
+              </label>
+              <label className="admin-dashboard__new-reservation-label">
+                Time:
+                <select
+                  value={newTime}
+                  onChange={(e) => setNewTime(e.target.value)}
+                  className=" admin-dashboard__new-reservation-input" 
+                  required
+                >
+                  <option className="admin-dashboard__option-time" value="">Select Time</option>
+                  {generateTimeSlots().map((slot, index) => (
+                    <option key={index} value={slot}>
+                      {slot}
+                    </option>
+                  ))}
+                </select>
+
+              </label>
+              <label className="admin-dashboard__new-reservation-label">
               Duration (minutes):
-              <input
-                type="number"
+              <select
                 value={newDuration}
                 onChange={(e) => setNewDuration(e.target.value)}
-                min="15"
-                step="15"
+                className=" admin-dashboard__new-reservation-input" 
                 required
-              />
-            </label>
-            <button
-              className="admin-dashboard__create-button"
-              onClick={handleCreateReservation}
-            >
-              Create Reservation
-            </button>
-          </div>
-
-          {/* Display Reservations */}
-          {isLoading ? (
-            <p>Loading reservations...</p>
-          ) : (
-            <table className="admin-dashboard__table">
-              <thead>
-                <tr>
-                  <th>Date</th>
-                  <th>Time</th>
-                  <th>Duration</th>
-                  <th>First Name</th>
-                  <th>Last Name</th>
-                  <th>Email</th>
-                  <th>Phone</th>
-                  <th>Final Price</th>
-                  <th>Actions</th>
-                </tr>
-              </thead>
-              <tbody>
-                {filteredReservations.length > 0 ? (
-                  filteredReservations.map((reservation) => (
-                    <tr key={reservation._id}>
-                      <td>{reservation.date}</td>
-                      <td>{reservation.time}</td>
-                      <td>{reservation.duration} mins</td>
-                      <td>{reservation.firstName}</td>
-                      <td>{reservation.lastName}</td>
-                      <td>{reservation.email}</td>
-                      <td>{reservation.phone}</td>
-                      <td>£{reservation.finalPrice}</td>
-                      <td>
-                        <button
-                          className="admin-dashboard__button"
-                          onClick={() => handleDeleteReservation(reservation._id)}
-                        >
-                          Delete
-                        </button>
-                      </td>
-                    </tr>
-                  ))
-                ) : (
-                  <tr>
-                    <td colSpan="9">No reservations found for the selected date.</td>
-                  </tr>
-                )}
-              </tbody>
-            </table>
-          )}
+              >
+                <option value="" className="admin-dashboard__option-time">Select Duration</option>
+                {durations.map((duration) => (
+                  <option key={duration.value} value={duration.value}>
+                    {duration.label}
+                  </option>
+                ))}
+              </select>
+              </label>
+              <label className="admin-dashboard__new-reservation-label">
+                First Name:
+                <input
+                  type="text"
+                  className="admin-dashboard__new-reservation-input"
+                  value={newFirstName}
+                  onChange={(e) => setNewFirstName(e.target.value)}
+                  required
+                />
+              </label>
+              <label className="admin-dashboard__new-reservation-label">
+                Last Name:
+                <input
+                  type="text"
+                  className="admin-dashboard__new-reservation-input"
+                  value={newLastName}
+                  onChange={(e) => setNewLastName(e.target.value)}
+                  required
+                />
+              </label>
+              <label className="admin-dashboard__new-reservation-label">
+                Email:
+                <input
+                  type="email"
+                  className="admin-dashboard__new-reservation-input"
+                  value={newEmail}
+                  onChange={(e) => setNewEmail(e.target.value)}
+                  required
+                />
+              </label>
+              <label className="admin-dashboard__new-reservation-label">
+                Phone:
+                <input
+                  type="text"
+                  className="admin-dashboard__new-reservation-input"
+                  value={newPhone}
+                  onChange={(e) => setNewPhone(e.target.value)}
+                  required
+                />
+              </label>
+              <button className="admin-dashboard__create-button" onClick={handleCreateReservation}>
+                Create Reservation
+              </button>
+            </div>
         </>
       ) : (
         <div className="admin-dashboard__coupons">
